Cosmetics for hard users
* Improved usage space on LuaLine
When in narrow screens (or terminals) or with vertically
split Neovim buffers, the Lua Line was a messy... so many
spaces for marginal information but so little for filename.

* Improved usage for old terminals or fonts users
Removed the need for icons and unusual characters on
LuaLine styling. This could break when not using an UTF-8
font or encoding.