Nim LSP Behavior improved
1. Removed shellscript wrapper/whatchdog for nimlsp

It makes nimlsp restart automatically after it ends,
but as the nvimlsp doenst provide a new handshake,
it doesn't serve the diagnostics.

Also there was a bug on my shellscript, that killed
the wrapper every new neovim instance. This make
impossible have two Neovim instances working correctly
with nimlsp. Now it works flawless (except by nimlsp
hangs)

2. Internalization of nimlsp watchdog

As nimlsp hangs sometimes (when you most need it) the
watchdog was moved to the filetype plugin configuration.
Now after every buffer write or buffer enter it checks
if nimlsp is alive. If not, it restarts it. I didn't
used extensively after this to check the efficacy, so
let a print to show on status when it is restarted.

Maybe some adjust (adding more frequent checks) be
needed, but not wanted it to be at first much exagerated.